# Another speed up to UI lag when moving lots of units.