After https://github.com/triplea-game/triplea/pull/5692, this is another set of optimizations. In this case, a paratrooper logic calculation is optimized by making that work O(n) rather than O(n^2), by moving it out of the inner "loop" - i.e. once for a list of units, instead processing a list of units for each unit.

Also refactors the code to add a helper method for duplicated code and fixes up formatting of some comments.

After the other PR, this change results in the following speed up:

Before the change:
Time to: Bastogne = 183 ms
Time to: Picardy = 51 ms
Time to: Compeigne = 56 ms
Time to: Aisne River = 58 ms

After the change:
Time to: Bastogne = 38 ms
Time to: Picardy = 30 ms
Time to: Compeigne = 32 ms
Time to: Aisne River = 36 ms

## Functional Changes
<!-- Put an X next any that apply -->
[] New map or map update
[] New Feature
[] Feature update or enhancement
[] Feature Removal
[X] Code Cleanup or refactor
[] Configuration Change
[X] Problem fix:  <!-- Link to bug issue or forum post here -->
[] Other:   <!-- Please specify -->

## Testing

[X] Manual testing done

Same as in https://github.com/triplea-game/triplea/pull/5692.